BUGFIX: QA-5830 Ability to handle paralell mergeability conflict
When there are multiple merges called at the same time, it leads to a
situation when github returns 405 MethodNotAllowed error becuase github
is checking the branch mergeability.

When we encounter this situation, we try to wait a bit (2 seconds for
now) and try to merge again.

It's also covered by simple tests using fake connection.

Change-Id: If490f3f3e69e09012638855297be0f79a838e91a